♻️ (Sidebar) - Split sidebar.js in separate files
